organize components, address changes
