+++ src/components/Rating/style.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+export const PoopContainer = styled.div`
+  display: flex;
+`;
+
+export const RatingPoop = styled.img`
+  width: ${(props) => props.width || "22px"};
+`;
