+++ src/components/Reply/style.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const ReplyContainer = styled.input`
+  background: #ececec;
+  filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
+  border-radius: 10px;
+  width: 80%;
+  gap: 10px;
+  margin-left: auto;
+  padding: 10px;
+  display: flex;
+
+  width: calc(80% - 20px);
+  border: none;
+  outline: none;
+
+  font-size: 14px;
+  text-align: left;
+`;
